Use separate handlers for disabled select option onMouseDown and onClick
Only in Microsoft Edge, clicking the a link in a disabled select option opens the href twice instead of once.

The disabled SelectOption's DropdownItem used to hook the same event handler for both "onMouseDown" and "onClick" events. A single click would trigger both of these events, thus calling the event handler twice and opening 2 windows/tabs instead of 1.

The fix is to only check for and follow links in one of the event handlers, not both.