Add default dateFormat to <NoteHeader /> to make it optional
This is for allowing dateFormat property of <NoteHeader /> to be
optinal.

The TypeScript type defition of fecha.format was incorrectly
making the 2nd parameter mask to be required, however, it would
actually be set to a default value 'ddd MMM DD YYYY HH:mm:ss' if not given.
https://github.com/taylorhakes/fecha/blob/8326090529b6f633c51750d6cf07625a8ef78f5b/fecha.js#L240
https://github.com/taylorhakes/fecha/blob/8326090529b6f633c51750d6cf07625a8ef78f5b/fecha.js#L213
As a result, not passing dateFormat does not cause any issue and
the TypeScript check and propType check is giving false alert.

We set the default dateFormat to fecha's default so that the
old behavior is kept and dateFormat became optional.