fix incorrect type declaration files
Many of the type declaration files were not semantically correct. This
was not caught when running the tslint script due to the skipLibCheck
flag in the tsconfig. This flag skips type checking on all type
declaration files. Due to the incorrect types, downstream projects have
to also turn on the skipLibCheck flag when using react-gears, even
if they do not use the modules with broken types.