Declare `summary` prop as optional in the `Paginator` component
The `Paginator` component has built-in handling for the absence of the
`summary` prop:
https://github.com/appfolio/react-gears/blob/1adbf65196bce045bc1596a602db6a4e3998a849/src/components/Paginator.tsx#L119
and there are numerous tests in `Paginator.spec.js` that do not pass the
`summary` prop:
https://github.com/appfolio/react-gears/blob/1adbf65196bce045bc1596a602db6a4e3998a849/test/components/Paginator.spec.js#L8-L189
Despite this, the `PaginatorProps` type added in
7037a741fe6850c4b9d2fe1bf908593571ef204f actually defined `summary` as a
required prop. This commit updates the type definition to correctly
declare it as optional.