Restore behavior of forwarding `Select`'s `name` prop to `SelectElement`
Prior to 198517b89ff6eef7d8dd9f08f02bc0475080e72e, the `name` prop
would be passed to `SelectElement` via the `{...props}` spread. But
after that commit, `name` is no longer part of the `props` rest
parameters, so it was no longer being passed to `SelectElement`. To fix
this, we explicitly pass `name={name}`. This is important because
without the `name` prop, the component won't render its underlying
hidden input field, as a result of this code:
https://github.com/HubSpot/react-select-plus/blob/v1.2.0/src/Select.js#L1148-L1149
```
	renderHiddenField (valueArray) {
		if (!this.props.name) return;
```