Update <FeatureBanner /> subtitle type
With the previous type, if a user wants the subtitle to have a link
and passed in a react component.

E.g
```jsx
const subtitle = <span>More info in <a href={url}>here</a>.</span>
<FeatureBanner
  alertText="New"
  subtitle={subtitle}
  title="Title"
>
...
```

The user gets prop type warning or TypeScript error while it works
perfectly.

The type of subtitle prop of <FeatureBanner /> was string, while
it can be anything that can be rendered, which is PropTypes.node,
React.ReactNode.

https://github.com/appfolio/react-gears/blob/c0e0eebbead311527be3b8b119fc94c3dc7f3a3d/src/components/FeatureBanner.js#L35