feat: set fade to false for feature banner
The fade property on the Reactstrap Alert component causes the alert
to fade in when it is rendered. This may make for a slightly smoother
experience when an alert pops up due to an error submitting a form, for
example. However, it is not as nice of an experience for static
components which are always rendered on the page, such as the feature
banner, since it causes them to take longer to render. It also causes
jarring behavior when swapping out some elements in the page while
leaving the feature banner unchanged.