Prevent Enter from submitting Forms
If ComboBox is used within a `form` element and there
is a `button` of type `submit`, whenever a user selects a value
using the `Enter` key, the value is selected then form is
automatically submitted.

This can be undesirable, especially if this is the first input
of the form and a user uses `Enter` to select a value.

In order to prevent this, we add a `preventDefault` call to the
`Enter` key section of the `onKeyDown` prop function. This will
prevent users from prematurely submitting forms while selecting
values in Combobox.