Makes city, state, postal col widths configurable
Since the column widths are based on the viewport size, if the
Address component is rendered within an element with a smaller
width, the city, state, and postal inputs can get squished and
unreadable in some cases. This makes it so you can override
those column widths to handle instances where this could be
a problem.