feat: add AllowWeakType util
We want to allow weak types in the case of typing generics for components like lists or tables in which we support optional properties like `key` or `expanded`. This will allow usage of the components in typescript without type errors if users choose not to provide any of the optional properties.