fix: allow conditional rendering of child elements
This fixes the case where null or false is passed as a child element to
CheckboxListInput or RadioInput. For instance:

    <CheckboxListInput>
      <FormChoice>Foo</FormChoice>
      {showBar && <FormChoice>Bar</FormChoice>}
    </CheckboxListInput>

Prior to this change, React.cloneElement would be called with null,
which would raise an exception.

The cast to any for the choice on RadioInput is necessary because
necessary because isValidElement is a type guard which casts its input
to ReactElement, which does not take "type" and "selected" as props.