feat: add API agnostic FeedbackButton
The react-apm feedback button contains assumptions about being called
from a property front end. This makes it so that it is not useful for
BFF use cases. This commit introduces a new FeedbackButton component
that includes all the functionality from the react-apm component except
for the default onSubmit actions. Instead the submit is performed via
the doSubmit callback property. This way BFF (and other) clients can use
the same feedback button experience but provide an alternate API
endpoint.