fix: work around typedef bug in react-use
React-use has a known issue in the typedef of PermissionDescriptor. The
skipLibCheck option will make it so we don't check the internal types of
npm dependencies. We may want to just keep this option on as it makes
the build faster and its rare to catch actionable issues with
skipLibCheck=false anyway.