fix: break long words in CollapsibleText
This fixes an issue where a long word, such as a URL, in the child of
the CollapsibleText component would overflow the boundary of the
container.

The `overflow-wrap: wrap-all` is preferred over `wordBreak: break-all`
because break-all would break even in the middle of short words if the
line character limit is reached in the middle of the word. It is
preferred over `wordBreak: break-word` because according to the MDN
page, that setting is deprecated in favor of overflow-wrap.